- some more safety checks in the ini-limit removal code - some code cleanup and minor bug fixes based on results from static code analysis - added naemfilter for the esp list - bsa changes are now stored automatically but delayed by up to 0.5 seconds (for performance reasons) - bugfix: buffer overrun when certain functions are called with empty file names - bugfix: reroute for the ini-limit fix was placed in the data segment - bugfix: mod list got mixed up when the mod directory was changed externally - bugfix: plugins.txt reroute on skyrim didn't work on winXP - bugfix: MO could become unresponsive if the tutorial script couldn't be interpreted

+++ src/disasm.cpp
@@ -475,7 +475,7 @@ void Disasm::ReadCommand(BYTE *pos)
 			_cmd_info.group = 42;
 			for (int i = 0; i < 17 && _cmd_info.group == 42; i++) {
 				if (optogroup[i][0] == _cmd_info.opcode) {
-					_cmd_info.group = optogroup[i][1];
+          _cmd_info.group = static_cast<short>(optogroup[i][1]);
 				}
 			}
 		}
